PR comments: - refactor useAppSelect, useAppDispatch imports - move Target component to Home components - refacto useHome to use redux action - refactor useEffect to useLayoutEffect in EmptyHomeSideBar - add home-action, home-reducer to home content - add home-reducer to root-reducer
