+++ src/hooks/useSelector.ts
@@ -1,3 +1,6 @@
 import { TypedUseSelectorHook, useSelector } from 'react-redux';
 import { RootState } from 'state/reducers/root-reducer';
-export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
+
+const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
+
+export default useAppSelector;
