+++ src/state/actions/home-actions.ts
@@ -1,0 +1,4 @@
+import { createAction } from '@reduxjs/toolkit';
+import { HomeContent } from 'pages/landing/Home/useHome';
+
+export const setHomeContent = createAction<HomeContent | undefined>('location/setHomeContent');
